rag/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
nltk==3.8.1
sentence-transformers==2.2.2
openai==1.12.0
pytest==8.0.0 